Move parser and ast to compile-time module
