--- Sources/Regex/AST/ASTTransform.swift
@@ -1,10 +1,0 @@
-extension CharacterClass {
-  public func withMatchLevel(
-    _ level: CharacterClass.MatchLevel
-  ) -> CharacterClass {
-      var cc = self
-      cc.matchLevel = level
-      return cc
-  }
-}
-
