+++ Sources/_MatchingEngine/Regex/AST/ASTEntity.swift
