+++ Sources/_MatchingEngine/Regex/Parse/LexicalAnalysis.swift
