+++ Sources/_MatchingEngine/Regex/Parse/SyntaxOptions.swift
